Split parameter view model onto a separate file
